requirements.txt
flask==3.0.2
python-dotenv==1.0.1
langchain-google-genai==0.0.11
langchain-core==0.1.27
psycopg2-binary==2.9.9 